Delete mystack.rs
Reasons for removing mystack.rs
1. Vector (a base class in Rust) already has the push and pop functions, and they would be more efficient than any implementation I would make. 
2. The compile time errors were getting worse: Here are methods I tried for this:
My initial attempt was to use a linked list method, clearly it failed, miserably, I have not become familiar enough with rust to create dynamic structs. 
My second attempt was using Arrays, and just keeping a tracker on what value in said array was the head. Then simply increment or decrement the head based on what function was called, and replace or retrieve the variable in the array.
One thing I have learned from this and how it effects coding style: Rust does allow for you to make code in "C-Style", but if this is done, it deems the whole goal of rust pointless. If you are trying to avoid the "C-style" rust is not nearly as friendly, I still need to get more familiar with dynamic structs. 
In the meantime, I will be using a Vector as a stack for the reverse polish notation calculator.