Modifications to functions.rs
Changes: 
A basic(ish) form of a reverse polish notation calculator has now been created. This took a long time. 
There are more things to be implemented as time goes on: 
I do plan on implementing a function that will convert reverse polish notation to infix expression, and with that I may be able to expand upon my simple arithmetic function to solve infix expressions that need an order of operations (I'll probably be using the Shunting-yard algorithm for that):
As you can see in the code: I have the rpn to infix commented out, and I have also attempted to create an external function to reverse a vector and return it. This is where I ran into my issues with the lifetime of variables.
The exact error I was getting is posted to the master branch as a PNG file so you can view it.   
I still need to learn more on this lifetime of variables thing, but the compiler errors do not make it clear, I eventually just gave up and moved all the stuff for reversing the vector into the