RPN to infix
So, I have added a function that converts rpn to infix notation:
Some things i have learned:
the pop() function does not return whatever value is in the stack. It instead returns an Option (std::option::Option) variable, in order to use this variable you have to unwrap it using unwrap().
Secondly: [].join(c: something) is a function in an array that returns all members of the array in a concatenated string seperating them by whatever parameter is passed into the function.
Finally: The errors, this code does not compile as the compiler is stating that the lifetime of the &str that is assigned by the join function (which I use the delimiter (&*) to convert to a &str), as always, the compiler errors are attached as a PNG.
That is all for this commit.