Finished complex calculator!!
Here, I have finished an implementation of the Shunting-Yard algorithm in rust to solve complex arithmetic (any arithmetic involving parentheses, order of operations, etc.).

This allowed me to use another provided data structure called a VecDeque, which is a double ended queue (I just used it as a queue). It allows you to use push_back() and pop() to create a queue. Very simple actually.

What I used the most here was the clone() function. Every variable can be cloned by using the .clone() function after it. By cloning, I mean getting the variable without the address, which isn't super useful here, but I can see how this can be VERY useful down the line. This is one of the parts that make Rust safer, and less likely to segfault (and the only part of said thing that I really understand at the moment).

Though I do not use the function, I have kept the rpn to infix function in there (still working on it) it still compiles, and when it runs it throws no errors, it just does not return the right value.

That is all for now.